Clean untagged images only if there are any.

+++ docker-compose/recipes/cleani.rb
@@ -1,4 +1,5 @@
 execute 'clean untagged images' do
     cwd '/docker-compose/'
     command '/usr/bin/docker rmi $(/usr/bin/docker images -q -f dangling=true)'
+    only_if '/usr/bin/docker images -q -f dangling=true'
 end
